backend/requirements.txt
# FastAPI и зависимости
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# База данных
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Аутентификация (упрощенная, без cryptography)
python-jose>=3.3.0
python-multipart>=0.0.6

# HTTP клиент для Planfix API
httpx>=0.26.0
requests>=2.31.0

# Утилиты
python-dotenv>=1.0.0
email-validator>=2.1.0


